Live - Debian: Changes for the Debian Live distro
Many of these changes were spur of the moment has to get done changes for the BKO release.
I'll happily admit, and take blame for the mess that they slightly create but this is whats
actually running in production right now so I figure getting it back into the tree and cleaning
it up after the fact is better than never committing.  These are almost identical to the changes
made to Ubuntu, though not quite exactly.

This is *NOT* complete.  There needs to be a libs/ directory that containts the contents of
/lib (no sub directories) to deal with the shards dependency of iscsiadm and iscsid to do
DNS resolution.  I'm reluctant to commit those binaries, as they are obivously binaries.
I'll be looking for a better way around this in the short term but for now that is *known* to
be broken unless you create the directory yourself.