Frontpage - Screenshot changes (thumbnails, css cleanups)
This provides a quick set of cleanups for the screenshots.  They were
the wrong size / ratio (making them look akward) and this sets them up
be be about right.  It doesn't look quite as nice from a space perspective
but they look better being the right ratio and all (images are clearer too)